backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
Pillow==10.2.0
numpy==1.26.3
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-cloud-vision==3.7.2
google-genai
